+++ tests/puppeteer/campaigns/functional/BO/customers/customers/02_CRUDCustomerInBO.js
@@ -8,8 +8,8 @@ const BOBasePage = require('@pages/BO/BObasePage');
 const LoginPage = require('@pages/BO/login');
 const DashboardPage = require('@pages/BO/dashboard');
 const CustomersPage = require('@pages/BO/customers');
-const AddCustomerPage = require('@pages/BO/addCustomer');
-const ViewCustomerPage = require('@pages/BO/viewCustomer');
+const AddCustomerPage = require('@pages/BO/customers/add');
+const ViewCustomerPage = require('@pages/BO/customers/view');
 const FOLoginPage = require('@pages/FO/login');
 const FOBasePage = require('@pages/FO/FObasePage');
 const CustomerFaker = require('@data/faker/customer');
