+++ tests/puppeteer/campaigns/functional/BO/international/taxes/04_taxOptionsForm.js
@@ -8,7 +8,7 @@ const taxOptions = require('@data/demo/taxOptions');
 const BOBasePage = require('@pages/BO/BObasePage');
 const LoginPage = require('@pages/BO/login');
 const DashboardPage = require('@pages/BO/dashboard');
-const TaxesPage = require('@pages/BO/taxes');
+const TaxesPage = require('@pages/BO/international/taxes');
 
 let browser;
 let page;
