+++ tests/puppeteer/campaigns/sanity/03_ordersBO/02_editOrder.js
@@ -8,7 +8,7 @@ const loginCommon = require('@commonTests/loginBO');
 const LoginPage = require('@pages/BO/login');
 const DashboardPage = require('@pages/BO/dashboard');
 const BOBasePage = require('@pages/BO/BObasePage');
-const OrderPage = require('@pages/BO/order');
+const OrderPage = require('@pages/BO/orders/view');
 const OrdersPage = require('@pages/BO/orders');
 const {Statuses} = require('@data/demo/orders');
