+++ tests/puppeteer/pages/BO/catalog/brands/view.js
