+++ tests/puppeteer/pages/BO/catalog/categories/add.js
