+++ tests/puppeteer/pages/BO/catalog/categories/index.js
