+++ tests/puppeteer/pages/BO/catalog/products/add.js
