+++ tests/puppeteer/pages/BO/catalog/products/index.js
