+++ tests/puppeteer/pages/BO/customers/add.js
