+++ tests/puppeteer/pages/BO/customers/index.js
