+++ tests/puppeteer/pages/BO/customers/view.js
