+++ tests/puppeteer/pages/BO/dashboard/index.js
