+++ tests/puppeteer/pages/BO/international/taxes/index.js
