+++ tests/puppeteer/pages/BO/modules/autoUpgrade/index.js
