+++ tests/puppeteer/pages/BO/modules/moduleCatalog/index.js
